# Add slow test stats and make some tests faster