Ref #8483. Adds a Karma plugin that displays 10 slowest tests after running the test suite:

```
2833ms GridLayer number of 256px tiles loaded in synchronous animated grid @800x600px Loads 290, unloads 275 tiles on MAD-TRD flyTo()
2813ms TileLayer number of kittens loaded Loads 290, unloads 275 kittens on MAD-TRD flyTo()                                          
809ms  Popup L.Map#openPopup moves the map over a long distance to the popup if it is not in the view (keepInView)                   
665ms  Map.ScrollWheelZoom changes the option 'wheelPxPerZoomLevel'                                                                  
606ms  Path #events can add a layer while being inside a moveend handler                                                             
365ms  Map.ScrollWheelZoom changes the option 'wheelDebounceTime'                                                                    
321ms  LineUtil #polylineCenter computes center of a small line and test it on every zoom                                            
310ms  Map.ScrollWheelZoom scrollWheelZoom: 'center'                                                                                 
306ms  Map.ScrollWheelZoom zooms out while firing 'wheel' event                                                                      
300ms  Map.ScrollWheelZoom zooms in while firing 'wheel' event     
```

Also makes a few of the unit tests faster, cutting down overall time by ~10 seconds.